+++ core/src/main/java/io/snabble/sdk/Snabble.java
@@ -135,6 +135,10 @@ public class Snabble {
         if (config.bundledMetadataAssetPath != null) {
             readMetadata();
             setupCompletionListener.onReady();
+
+            if (config.loadActiveShops) {
+                loadActiveShops();
+            }
         } else {
             metadataDownloader.loadAsync(new Downloader.Callback() {
                 @Override
@@ -151,6 +155,10 @@ public class Snabble {
                     }
 
                     setupCompletionListener.onReady();
+
+                    if (config.loadActiveShops) {
+                        loadActiveShops();
+                    }
                 }
 
                 @Override
@@ -479,13 +487,28 @@ public class Snabble {
             protected void onDataLoaded(boolean wasStillValid) {
                 if (!wasStillValid) {
                     readMetadata();
+                    notifyMetadataUpdated();
+                }
 
-                    for (OnMetadataUpdateListener listener : onMetaDataUpdateListeners) {
-                        listener.onMetaDataUpdated();
-                    }
+                if (config.loadActiveShops) {
+                    loadActiveShops();
                 }
             }
         });
+    }
+
+    private void loadActiveShops() {
+        for (Project project : projects) {
+            project.loadActiveShops(() -> {
+                notifyMetadataUpdated();
+            });
+        }
+    }
+
+    private void notifyMetadataUpdated() {
+        for (OnMetadataUpdateListener listener : onMetaDataUpdateListeners) {
+            listener.onMetaDataUpdated();
+        }
     }
 
     private void checkCartTimeouts() {
@@ -753,5 +776,10 @@ public class Snabble {
 
         /** Vibrate while adding a product to the cart, by default false */
         public boolean vibrateToConfirmCartFilled = false;
+
+        /** Set to true, to load shops that are marked as pre launch
+         *  and are not part of the original metadata in the backend
+         *  (for example for testing shops in production before a go-live) **/
+        public boolean loadActiveShops = false;
     }
 }